requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# AI/ML Libraries
openai>=1.10.0
anthropic>=0.18.1
google-generativeai>=0.3.0
groq>=0.4.0
sentence-transformers>=2.2.0

# Vector Database
qdrant-client>=1.7.3

# Audio Processing (for Voice Karaoke)
# openai-whisper>=20231117  # Uncomment when needed - large download
# pydub>=0.25.1
# librosa>=0.10.1
# soundfile>=0.12.1

# PDF Processing
PyPDF2>=3.0.1

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.26.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

# Logging & Monitoring
python-json-logger>=2.0.7
